+++ crates/rbuilder-utils/Cargo.toml
@@ -9,6 +9,8 @@ repository.workspace = true
 exclude.workspace = true
 
 [dependencies]
+reth-tasks = { git = "https://github.com/paradigmxyz/reth", rev = "9c30bf7af5e0d45deaf5917375c9922c16654b28" }
+
 # misc
 derivative.workspace = true
 integer-encoding = "4.0.0"
