+++ crates/rbuilder-utils/src/clickhouse/mod.rs
@@ -1,3 +1,5 @@
+pub mod backup;
+pub mod indexer;
 use serde::{Deserialize, Serialize};
 
 /// Equilalent of `clickhouse::inserter::Quantities` with more traits derived.
