+++ crates/rbuilder-utils/src/lib.rs
@@ -1,5 +1,7 @@
+pub mod backoff;
 pub mod clickhouse;
-pub mod clickhouse_with_backup;
 pub mod format;
 pub mod metrics;
-pub mod tokio;
+pub mod tasks {
+    pub use reth_tasks::*;
+}
